requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
twilio==8.10.0
langgraph==0.0.49
langchain==0.1.13
python-dotenv==1.0.0
pydantic==2.5.0
